Add UserInput Interface, App.Config and UserInput class
Created a UserInput interface, just to test interfaces a bit. Implemented the interface in the UserInput class. Built code in UserInput class to Try to get user input string, a method to check that the user input string is part of a correct choice, and also a method to rip a string of formatting - which for now just removes all spaces and makes it all lower case.