Add MenuModel and TableVisualisationEngine
MenuModel will be the model for creating menus. It contains a string MenuTitle and a MenuItems which is a List of List of objects. This model will be used in conjunction with the TableVisualisationEngine to print out the menu to the user. TableVisualisationEngine will use the MenuTitle as the title to the table, and the MenuItems as the row and column data.