Fixed an input issue
There was an issue where I take all spaces and capitalization away from the user input and the available options, and I returned this "deformatted" string as the output. This could cause issues with calling from the API. So now userInput will return a string that comes directly from the API which matches the "deformatted" user input.